Make Station it's own entity.
This allows us to let OrderedStation extend it, and share some common
code.
