+++ entities/OrderedStation.php
@@ -6,14 +6,7 @@ use Respect\Relational\Mapper;
 use Routelandia\DB;
 
 /**
- * Represents a single Station
- *
- * == Notes about Stations in the Portal Database ==
- * Station's have ID's grouped into thousands:
- *   1000's: Inductive loop detectors
- *   2000's: HOV lane detectors in vancouver. (Or maybe elsewhere later)
- *   3000's: HD Radar detectors.
- *   5000's: Onramp detectors.
+ * Represents a single row in the orderedStations view.
  *
  * An OrderedStation is much like a
  * station, except it's coming in via
@@ -27,7 +20,7 @@ use Routelandia\DB;
  * specific highway. (The ordering doesn't make much
  * sense when you select all of them)
  */
-class OrderedStation {
+class OrderedStation extends Station {
 
   public $stationid;
   public $upstream;
@@ -138,29 +131,6 @@ class OrderedStation {
    * Currently will only return a single onramp, but the possibility is there...
    */
   public static function fetchRelatedOnramps($id) {
-    $onRamp = DB::instance()->stations(array('stationid='=>$id))->fetch();
+    return Station::fetch($id);
   }
-
-
-  /**
-   * Returns the ID of the onramp detector related to this station.
-   *
-   * Related onramps are detected by having the "same" ID in the 5000 range.
-   * i.e. Station 1037 should have an onramp 5037, if such an onramp exists.
-   * Note that onramps aren't useful for speed, because they're just a single loop.
-   * @return int -1 if not possible, otherwise the ID that the onramp *should* be.
-   */
-  public static function calculateRelatedOnrampID($tid) {
-    if($tid >= 1000 && $tid < 4000) {
-      // First we strip it down to the base ID. (not in the thousands range.)
-      while($tid > 1000) {
-        $tid = $tid-1000;
-      }
-
-      return $tid+5000;
-    } else {
-      return null;
-    }
-  }
-
 }
